把测心率部分加上去了。美化了UI。 by 老爷爷

+++ app/src/main/java/com/bigboss/heartrate/util/CameraHelper.java
@@ -82,15 +82,15 @@ public class CameraHelper {
         }
     }
 
-    static public int decodeYUV420StoGreyRGBandGetAvgGrey(int[] rgb, byte[] yuv420sp, int width, int height) {
-        int sumofGrey = 0;
+    static public float decodeYUV420StoGreyRGBandGetAvgGrey(int[] rgb, byte[] yuv420sp, int width, int height) {
+        float sumofGrey = 0;
         int frameSize = width*height;
         for (int i = 0; i < frameSize; i++) {
             int grey = yuv420sp[i] & 0xff;
             sumofGrey += grey;
             rgb[i] = 0xFF000000 | (grey * 0x00010101);
         }
-        return sumofGrey/frameSize;
+        return (float)sumofGrey/(float)frameSize;
     }
 
     public static void setAutoFocusMode(Camera camera) {
